Fix hourly sync crash: unify worker event loop across all task modules
Four Celery task modules (sync, workflows, monitoring, bulk_operations)
each maintained their own `_worker_loop` / `asyncio.run()`, creating
multiple event loops in the same worker process. When asyncpg connections
created on one loop were checked out by a task on another loop, the worker
crashed with `RuntimeError: Future attached to a different loop` — silently
killing hourly syncs for hours at a time.

Extracts a single shared `run_async()` into `workers/run_async.py` so all
task modules share one event loop per worker process.

Made-with: Cursor
