+++ backend/workers/tasks/bulk_operations.py
@@ -32,6 +32,7 @@ from celery import group as celery_group
 from sqlalchemy import select, func as sa_func, text
 
 from workers.celery_app import celery_app
+from workers.run_async import run_async
 
 logger = logging.getLogger(__name__)
 
@@ -48,26 +49,6 @@ MAX_BULK_ITEMS: int = 50_000
 # Redis key helpers
 def _redis_key(operation_id: str, field: str) -> str:
     return f"bulk_op:{operation_id}:{field}"
-
-
-_worker_loop: asyncio.AbstractEventLoop | None = None
-
-
-def run_async(coro: Any) -> Any:
-    """Run an async function in a sync Celery task context.
-
-    Reuses a single event loop per worker process so that asyncpg connections
-    remain valid across task invocations.
-    """
-    global _worker_loop
-
-    if _worker_loop is None or _worker_loop.is_closed():
-        from models.database import dispose_engine
-        dispose_engine()
-        _worker_loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(_worker_loop)
-
-    return _worker_loop.run_until_complete(coro)
 
 
 async def _mark_operation_failed(operation_id: str, organization_id: str, error: str) -> None:
