+++ backend/workers/tasks/sync.py
@@ -14,13 +14,13 @@ _backend_dir = Path(__file__).resolve().parent.parent.parent
 if str(_backend_dir) not in sys.path:
     sys.path.insert(0, str(_backend_dir))
 
-import asyncio
 import logging
 from datetime import datetime, timedelta
 from typing import Any
 from uuid import UUID
 
 from workers.celery_app import celery_app
+from workers.run_async import run_async
 
 logger = logging.getLogger(__name__)
 
@@ -30,26 +30,6 @@ PROVIDER_SYNC_INTERVALS: dict[str, timedelta] = {
     "google_drive": timedelta(minutes=30),
 }
 DEFAULT_SYNC_INTERVAL: timedelta = timedelta(hours=1)
-
-
-_worker_loop: asyncio.AbstractEventLoop | None = None
-
-
-def run_async(coro: Any) -> Any:
-    """Run an async function in a sync context (for Celery tasks).
-
-    Reuses a single event loop per worker process so that asyncpg connections
-    remain valid across task invocations.
-    """
-    global _worker_loop
-
-    if _worker_loop is None or _worker_loop.is_closed():
-        from models.database import dispose_engine
-        dispose_engine()
-        _worker_loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(_worker_loop)
-
-    return _worker_loop.run_until_complete(coro)
 
 
 async def _sync_integration(
